[Test] Mamba2: Trim the batch-invariant tests to one case per property
Fold the four test files into three and 37 cases into 11:

- The kernel-level schedule now runs two layers that share one metadata
  object with different state slots on the paged layout, which is what the
  model runner does for the KV cache groups of a hybrid model; it replaces
  the separate shared-metadata regression test and the layout and seed
  parametrization.
- The state layout, startup checks and the metadata of decode and prefill
  rows live in one file; the randomized metadata property test is dropped
  since the kernel-level schedule exercises the same layouts end to end.
